# go-server/go/model_error_not_found.go
/*
 * Loan Service API
 *
 * This is a simple server to create and get information about Loan service.
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type ErrorNotFound struct {

	Status int32 `json:"status,omitempty"`

	Message string `json:"message,omitempty"`
}