# go-server/go/model_track_input.go
/*
 * Loan Service API
 *
 * This is a simple server to create and get information about Loan service.
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type TrackInput struct {

	// date in format YYYY-MM-DD
	Date string `json:"date,omitempty"`
}