# go-server/go/model_track_response.go
/*
 * Loan Service API
 *
 * This is a simple server to create and get information about Loan service.
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type TrackResponse struct {

	Count int32 `json:"count,omitempty"`

	Summary string `json:"summary,omitempty"`

	Var7DayAverage string `json:"7-day-average,omitempty"`
}